requirements.txt
aiofiles==23.2.1
async-timeout==4.0.3
asyncpg==0.28.0
bcrypt==4.0.1
blinker==1.5
boto3==1.28.24
botocore==1.31.24
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.14.4
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
priority==2.0.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
quart==0.18.4
quart-cors==0.6.0
Quart-JWT-Extended==0.1.0
quart-rate-limiter==0.8.0
redis==4.6.0
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
urllib3==1.26.16
Werkzeug==2.3.6
wsproto==1.2.0
async_stripe==5.5.0
